backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1
